+++ s3os/__init__.py
@@ -1,2 +1,6 @@
-from s3os import main
-os = main.os()
+from s3os import os
+
+from s3os import shutil
+os = os.OS()
+
+shutil = shutil.Shutil()
