+++ s3os/os.py
@@ -1,39 +1,23 @@
 import boto3
+from s3os import path
 
-
-class Path():
-    def __init__(self, os_obj):
-        self.os_obj = os_obj
-    
-    def exists(self):
-        raise Exception("Not implemented in this version")
-    
-    def dirname(self):
-        raise Exception("Not implemented in this version")
-        
-    def basename(self):
-        raise Exception("Not implemented in this version")
-
-    def is_dir(self):
-        raise Exception("Not implemented in this version")
-    
-    def is_file(self):
-        raise Exception("Not implemented in this version")
-
-class os():
+class OS():
     def __init__(self):
-        self.path = Path(self)
+        self.path = path.Path(self)
 
     def authorize(self, bucket, aws_access_key_id=None, aws_secret_access_key=None):
         self.s3 = boto3.resource('s3', aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_access_key)
         self.bucket = self.s3.Bucket(bucket)
-    
-    def listdir(self, directory):
+
+    def __listdir(self, directory):
         if directory.startswith("\\") or directory.startswith("/"):
             directory = directory[1:]
-        print(directory)
-        for bucket_object in self.bucket.objects.filter(Prefix=directory, Delimiter='/'):
-            print(bucket_object.key)
+#         for bucket_object in self.bucket.objects.filter(Prefix=directory, Delimiter='/'):
+        for bucket_object in self.bucket.objects.filter(Prefix=directory):
+            yield bucket_object.key
+
+    def listdir(self, directory):        
+        return list(self.__listdir(directory))
 
     def walk(self):
         raise Exception("Not implemented in this version")
@@ -54,13 +38,4 @@ class os():
         raise Exception("Not implemented in this version")
         
     def unlink(self):
-        raise Exception("Not implemented in this version")
-        
-def shutil():
-    def authorize(self, bucket, aws_access_key_id=None, aws_secret_access_key=None):
-        self.s3 = boto3.resource('s3', aws_access_key_id=aws_access_key_id, aws_secret_access_key=aws_secret_access_key)
-        self.bucket = self.s3.Bucket(bucket)
-
-    def copy(self, src, dst):
-        ''' Not a OS function, but a shutil function adding to improve usage'''
         raise Exception("Not implemented in this version")
